requirements.txt
asgiref==3.6.0
contourpy==1.0.6
cycler==0.11.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.4
django-import-export==3.0.2
et-xmlfile==1.1.0
fonttools==4.38.0
kiwisolver==1.4.4
MarkupPy==1.14
matplotlib==3.6.2
numpy==1.24.1
odfpy==1.4.1
openpyxl==3.0.10
packaging==22.0
Pillow==9.3.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tablib==3.3.0
tzdata==2022.7
xlrd==2.0.1
xlwt==1.3.0
